pastelLuna/requirements.txt
django
mysqlclient
django-mathfilters
pillow
djangorestframework
bcrypt
password-strength  
whitenoise
python-magic-bin
